backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
openai>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.28
alembic>=1.13.1
numpy>=1.26.4
faiss-cpu>=1.8.0
sentence-transformers>=2.5.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.26.0 